Lab 7: Added submission information, changed to include flashcards.html
Before, it was written as if flashcards.html was not provided, so there was
extra information to help the students write it. Now, it has been edited with
the assumption that they will use this file and just add an upload capability.
Has not yet been implemented on own machine.